Implement user authentication tests and update auth system
Expanded the authentication system in the AuthController to include user registration, login and logout functionality. Built a test suite for these features in AuthTest.php with successful and failure cases. Heavily commented Open API descriptors were added to the controller and bundled in the stored JSON documentation. An .env.testing file was also created for establishing test specific environment variables, and the command `php artisan test` was added to the package.json file.